+++ R/mat2latex.R
@@ -1,11 +1,15 @@
 #' Convert an R Matrix Object to LaTeX Code
-#' @description This function converts an R matrix object into LaTeX code for rendering it in a LaTeX environment. It is intended for use in Rmarkdown,
+#' @description This function converts an R matrix object into LaTeX code for
+#'  rendering it in a LaTeX environment. It is intended for use in Rmarkdown,
 #' as a convenience for translating calculations done in R code chunks into LaTeX.
 #' @param m An R matrix object of any dimension
-#' @param sink Logical determining output mode. If FALSE, the default, mat2latex prints the LaTeX code directly to console. If TRUE,
+#' @param sink Logical determining output mode. If FALSE, the default, mat2latex
+#'  prints the LaTeX code directly to console. If TRUE,
 #' it returns the code without printing, a la R's sink function.
-#' @return Latex code for printing the matrix. To render the code in an RMarkdown document, call the function in a chunk with  the `results = "asis"` option.
-#' Alternately, set sink to FALSE, store the output in an object, and reference the object in an R code chunk.
+#' @return Latex code for printing the matrix. To render the code in an RMarkdown
+#'  document, call the function in a chunk with  the `results = "asis"` option.
+#' Alternately, set sink to FALSE, store the output in an object, and reference
+#'  the object in an R code chunk.
 #' @export
 #'
 #' @examples #Large matrices are not difficult to process.